backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Firebase (includes Firestore)
firebase-admin>=6.0.0

# Google Generative AI
google-generativeai>=0.3.0
